Fix warnings
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@11509 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ exp/include/Galois/Runtime/ParallelWorkInline.h
@@ -145,13 +145,13 @@ class dChunkedMaster : private boost::noncopyable {
     if (r)
       return r;
     
-    for (int i = id.pid + 1; i < (int) Q.size(); ++i) {
+    for (unsigned int i = id.pid + 1; i < Q.size(); ++i) {
       r = popChunkByID(i);
       if (r) 
 	return r;
     }
 
-    for (int i = 0; i < id.pid; ++i) {
+    for (unsigned int i = 0; i < id.pid; ++i) {
       r = popChunkByID(i);
       if (r)
 	return r;
